requirements.txt
Django>=5.2.8
psycopg2-binary>=2.9.11
stripe>=14.0.1
Pillow>=12.0.0
django-crispy-forms>=2.5
crispy-bootstrap5>=2025.6
python-decouple>=3.8
django-extensions>=4.1
whitenoise>=6.6.0
gunicorn>=21.2.0
dj-database-url>=2.1.0

